requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.2
numpy==1.26.2
scikit-learn==1.8.0
scipy==1.11.3
redis==5.0.5
prometheus-client==0.18.0
python-dotenv==1.0.0
joblib==1.3.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
apscheduler==3.10.1
redlock-py==1.0.8